Move twist & taper fields to separate module.
